@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,28 +22,35 @@ dependencies:
   fit_kit: ^1.1.1
   health: ^1.0.2
   path: ^1.6.4
-  googleapis: ^0.54.0
-  google_sign_in: ^4.4.4
+  googleapis: ^0.55.0
+  google_sign_in: ^4.5.1
 
   charts_flutter: ^0.9.0
   intl: ^0.16.1
   translator: ^0.1.4
   flutter_cupertino_localizations: ^1.0.1
-  flutter_cupertino_date_picker: ^1.0.26
+  flutter_cupertino_date_picker: ^1.0.26+2
   font_awesome_flutter: ^8.8.1
-  firebase_auth: ^0.16.0
+  firebase_auth: ^0.16.1
   audioplayers: ^0.15.1
-  cupertino_icons: ^0.1.2
-  sqflite: ^1.3.0
-  floor: ^0.12.0
-  provider:  ^4.0.5+1
-  moor_flutter: ^2.1.1
-  shared_preferences: ^0.5.7+1
+  cupertino_icons: ^0.1.3
+  sqflite: ^1.3.1
+  provider:  ^4.3.1
+  moor_flutter: ^3.1.0
+  shared_preferences: ^0.5.8
   get_it: ^4.0.2
+  moor_db_viewer: ^2.0.3
+  connectivity: ^0.4.9
+  hive: ^1.4.1+1
+  hive_flutter: ^0.3.0+2
 
+  analyzer_plugin: ^0.2.4
+dependency_overrides:
+  dartx: ^0.3.0
 dev_dependencies:
-  moor_generator: ^2.4.0
-  build_runner: ^1.0.0
+  moor_generator: ^3.2.0
+  build_runner: ^1.10.0
+  hive_generator: ^0.7.0+2
   flutter_test:
     sdk: flutter
   
