@@ -1,5 +1,6 @@
-const String Brazos = "Brazos"; 
-const String Piernas = "Piernas"; 
-const String Tronco = "Tronco"; 
-const String Espalda = "Espalda"; 
+const String Brazos = "Brazos";
+const String Piernas = "Piernas";
+const String Tronco = "Tronco";
+const String Espalda = "Espalda";
 const String Cuello = "Cuello";
+const List<String> Parts = ["Brazos", "Piernas", "Tronco", "Espalda", "Cuello"];
