@@ -8,12 +8,14 @@ import 'package:google_sign_in/google_sign_in.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class LoginState with ChangeNotifier {
-  GoogleSignIn _googleSingIn = GoogleSignIn(scopes: ['https://www.googleapis.com/auth/drive'],);
+  GoogleSignIn _googleSingIn = GoogleSignIn(
+    scopes: ['https://www.googleapis.com/auth/drive'],
+  );
   GoogleSignInAccount _googleAcc;
   SharedPreferences _prefs;
   Map<String, String> header;
   final _auth = FirebaseAuth.instance;
-  var  _idToken;
+  var _idToken;
   bool _fecha_introducida = false;
   bool _logedIn = false;
   bool _loading = true;
@@ -23,7 +25,7 @@ class LoginState with ChangeNotifier {
   bool isLoading() => _loading;
   DateTime _date;
   bool isLogedIn() => _logedIn;
-  String getidToken()=> _idToken;
+  String getidToken() => _idToken;
   FirebaseUser currentUser() => _user;
   bool getFecha() => _fecha_introducida;
   String getId() => _id;
@@ -33,31 +35,34 @@ class LoginState with ChangeNotifier {
   }
 
   void setDate(DateTime date) => _date = date;
-  void setLogedIn(){
+  void setLogedIn() {
     _logedIn = true;
     _loading = false;
-    _fecha_introducida = true;
+    _fecha_introducida = false;
     _correctRestore = true;
     _prefs.setBool('completeLogin', true);
     notifyListeners();
   }
-  void setLogedInNoRestore(){
-    _logedIn = true;
+
+  void setLogedInNoRestore() {
+    _logedIn = false;
     _loading = false;
     _fecha_introducida = true;
-    _prefs.setBool('completeLogin', true);
+    _prefs.setBool('completeLogin', false);
     notifyListeners();
   }
+
   DateTime getFNacimiento() => _date;
   void setFecahIntroducida() {
     _fecha_introducida = true;
     notifyListeners();
   }
-  
-  void setLoading() { 
+
+  void setLoading() {
     _loading = true;
     notifyListeners();
   }
+
   void insertarFecha(DateTime dateTime) async {
     _fecha_introducida = true;
     _date = dateTime;
@@ -65,7 +70,6 @@ class LoginState with ChangeNotifier {
   }
 
   void insert(AppDatabase database) async {
-    
     UsuarioData usuario = UsuarioData(
         id: _id,
         nombre: _user.displayName,
@@ -74,12 +78,11 @@ class LoginState with ChangeNotifier {
         email: _user.email);
     database.usuarioDAO.insertUser(usuario);
     _prefs.setBool('completeLogin', true);
-    
+
     _logedIn = true;
 
     notifyListeners();
   }
-
 
   void login(AppDatabase database, BuildContext context) async {
     _prefs.setBool('completeLogin', false);
@@ -88,42 +91,42 @@ class LoginState with ChangeNotifier {
     try {
       _user = await _loginGoogle();
 
+      _id = _user.uid;
+      //_date = await _getBDay();
 
-    _id = _user.uid;
-    //_date = await _getBDay();
+      print("Birthday from " + _user.displayName + " " + _date.toString());
+      _loading = false;
+      if (_user != null) {
+        _prefs.setBool('isLoggedIn', true);
+        Backup b =
+            Backup(header: header, id: _id, database: database, state: this);
 
-    print("Birthday from " + _user.displayName + " " + _date.toString());
-    _loading = false;
-    if (_user != null) {
-      
-      _prefs.setBool('isLoggedIn', true);
-      Backup b = Backup(header: header, id: _id, database:database , state: this);
-      
-      var z = await database.usuarioDAO.getUser(_id);
-      
-      
-      
-      if ( z == null){
-         
-        await b.getDataFronDrive();
-        _fecha_introducida = true;
+        var z = await database.usuarioDAO.getUser(_id);
+
+        if (z == null) {
+          try {
+            await b.getDataFronDrive();
+            _fecha_introducida = true;
+            notifyListeners();
+          } catch (e) {
+            _fecha_introducida = false;
+            notifyListeners();
+          }
+        } else {
+          _prefs.setBool('completeLogin', true);
+          _fecha_introducida = false;
+          _logedIn = true;
+        }
+
+        notifyListeners();
+      } else {
+        _logedIn = false;
+        notifyListeners();
       }
-      else {
-         _prefs.setBool('completeLogin', true);
-        _fecha_introducida = false;
-        _logedIn = true;
-      }
-
-      notifyListeners();
-    } else {
-     
-      _logedIn = false;
-      notifyListeners();
-    }
-  }catch(e){
+    } catch (e) {
       _loading = false;
       notifyListeners();
-      throw(e);
+      throw (e);
       //login(database, context);
       //print(e);
     }
@@ -138,16 +141,15 @@ class LoginState with ChangeNotifier {
     notifyListeners();
   }
 
-
-  Future<FirebaseUser> _loginGoogle() async{
-     
-   final GoogleSignInAccount googleSignInAccount =  
-         await _googleSingIn.signIn();  
-     return _handleSignIn(googleSignInAccount);  
+  Future<FirebaseUser> _loginGoogle() async {
+    final GoogleSignInAccount googleSignInAccount =
+        await _googleSingIn.signIn();
+    return _handleSignIn(googleSignInAccount);
   }
+
   Future<FirebaseUser> _handleSignIn(GoogleSignInAccount account) async {
     //account = await _googleSingIn.signIn();
-    
+
     //var hs = await account.authHeaders;
     //var h = hs.values.first;
     //_prefs.setString('Header', h);
@@ -160,25 +162,24 @@ class LoginState with ChangeNotifier {
         (await _auth.signInWithCredential(credential)).user;
     print("signed in " + user.displayName);
     header = await account.authHeaders;
-    
 
     return user;
   }
 
   void loginState() async {
-    
     _prefs = await SharedPreferences.getInstance();
     if (_prefs.containsKey('isLoggedIn') && _prefs.getBool('completeLogin')) {
-       //header = _prefs.getString('Header');
-      
-      await _googleSingIn.signInSilently().whenComplete(() => () {}); 
-      _googleAcc =  _googleSingIn.currentUser;
+      //header = _prefs.getString('Header');
+
+      await _googleSingIn.signInSilently();
+      _googleAcc = _googleSingIn.currentUser;
+
       header = await _googleAcc.authHeaders;
-      
+
       _user = await _auth.currentUser();
-      
+
       //header = await _googleSingIn.currentUser.authHeaders;
-      
+
       _logedIn = _user != null;
       _id = _user.uid;
       _loading = false;
@@ -189,7 +190,4 @@ class LoginState with ChangeNotifier {
       notifyListeners();
     }
   }
-  
 }
-
-
