@@ -1,31 +1,26 @@
-
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-//import 'package:flutterapp/RouteManager.dart';
-import 'package:flutterapp/Alertas/Alertas.dart';
-import 'package:flutterapp/NavigationTools/locator.dart';
-import 'package:flutterapp/NavigationTools/navigator_service.dart';
-
-import 'package:flutterapp/ejercicios/AppTimer.dart';
-import 'package:flutterapp/ejercicios/AppRepCount.dart';
-import 'package:flutterapp/ejercicios/Ejercicio.dart';
-import 'package:flutterapp/ejercicios/EjercicioRepeticiones.dart';
-import 'package:flutterapp/ejercicios/EjercicioTiempo.dart';
-
-//import 'package:flutterapp/pulsera/datosRitmoTR/alertaRitmo.dart';
-import 'package:flutterapp/pulsera/datosRitmoTR/sacaDatosRitmoCardiaco.dart';
-
-import 'package:flutterapp/Data/moor_database.dart';
-import 'package:provider/provider.dart';
-import 'package:flutterapp/Registro/SignUpState.dart';
 import 'dart:async';
-import 'package:flutterapp/NavigationTools/routes_path.dart' as route;
-
-import 'EjercicioRepeticiones.dart';
-import 'EjerciciosState.dart';
 
 import 'package:audioplayers/audio_cache.dart';
 import 'package:audioplayers/audioplayers.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutterapp/Alertas/Alertas.dart';
+import 'package:flutterapp/Data/HiveData/RecomendadorList/RecomList.dart';
+import 'package:flutterapp/Data/moor_database.dart';
+import 'package:flutterapp/NavigationTools/locator.dart';
+import 'package:flutterapp/NavigationTools/navigator_service.dart';
+import 'package:flutterapp/Registro/SignUpState.dart';
+import 'package:flutterapp/ejercicios/AppRepCount.dart';
+import 'package:flutterapp/ejercicios/AppTimer.dart';
+import 'package:flutterapp/ejercicios/Ejercicio.dart';
+import 'package:flutterapp/ejercicios/EjercicioRepeticiones.dart';
+import 'package:flutterapp/ejercicios/EjercicioTiempo.dart';
+import 'package:flutterapp/pulsera/datosRitmoTR/sacaDatosRitmoCardiaco.dart';
+import 'package:hive/hive.dart';
+import 'package:provider/provider.dart';
+
+import 'EjercicioRepeticiones.dart';
+import 'EjerciciosState.dart';
 
 class BuildEjercicio extends StatefulWidget {
   Ejercicio ejercicio;
@@ -40,26 +35,31 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
   AudioCache _audioCache;
   bool sacaaudio = false;
   int cont = 0;
-
+  Box recomListBox;
   @override
   void initState() {
     super.initState();
     //int edad = sacarEdad(); no va por el future
+
     setAuxExceso(false);
-    _audioCache = AudioCache(prefix: "audio/", fixedPlayer: AudioPlayer()..setReleaseMode(ReleaseMode.STOP));
-    timer = Timer.periodic(Duration(seconds: 3), (Timer t){ // En principio cada 3 seg
-      if(!getAuxExceso()) { // Si no salta la alerta
+    _audioCache = AudioCache(
+        prefix: "audio/",
+        fixedPlayer: AudioPlayer()..setReleaseMode(ReleaseMode.STOP));
+    timer = Timer.periodic(Duration(seconds: 3), (Timer t) {
+      // En principio cada 3 seg
+      if (!getAuxExceso()) {
+        // Si no salta la alerta
         sacarEdad();
         cont = 0;
-      }
-      else{ // Si salta
+      } else {
+        // Si salta
         cont++;
         print(cont);
-        if(!sacaaudio) {
+        if (!sacaaudio) {
           _audioCache.play("beep-09.mp3");
           sacaaudio = true;
         }
-        if(cont == 4) {
+        if (cont == 4) {
           setAuxExceso(false);
           sacaaudio = false;
         }
@@ -67,39 +67,34 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
     });
   }
 
-  void botonHecho(BuildContext context,EjercicioState ejstate){
+  void botonHecho(BuildContext context, EjercicioState ejstate) {
     Alerts alerta = Alerts(
       context: context,
       title: "Ejercicio Completado",
       message: "¿Que datos deseas usar?",
-
       firstButtonText: "Datos del Ejercicio",
       secondButtonText: "Datos del Usuario",
       thirdButtonText: "cancelar",
-      fun1:(){ 
+      fun1: () {
         addEjercicio(context, ejstate, widget.ejercicio, true);
 
         _navigationService.goBack();
-        },
-      fun2:() {
-        addEjercicioBase(context,ejstate);
-        _navigationService.goBack();
-
       },
-      fun3:()=> _navigationService.goBack(),
-
-
+      fun2: () {
+        addEjercicioBase(context, ejstate);
+        _navigationService.goBack();
+      },
+      fun3: () => _navigationService.goBack(),
     );
     alerta.showAlertDialog3();
     return;
   }
-  addEjercicioBase(BuildContext context,EjercicioState ejstate){
+
+  addEjercicioBase(BuildContext context, EjercicioState ejstate) {
     dynamic aux = widget.ejercicio;
-    if(aux is EjercicioTiempo){
+    if (aux is EjercicioTiempo) {
       aux.setTime(ejstate.getTime);
-      
-    }else{
-      
+    } else {
       aux.setReps(ejstate.getReps);
       aux.setSeries(ejstate.getSeries);
     }
@@ -111,18 +106,21 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
     timer?.cancel();
     super.dispose();
   }
-  void addEjercicio(BuildContext context, EjercicioState ejstate, Ejercicio aux, bool ok) async {
+
+  void addEjercicio(BuildContext context, EjercicioState ejstate, Ejercicio aux,
+      bool ok) async {
     final database = Provider.of<AppDatabase>(context, listen: false);
     var state = Provider.of<LoginState>(context, listen: false);
+    var r2;
     String uid = state.getId();
     if (aux is EjercicioTiempo) {
       EjercicioTiempo ej = aux;
       EjercicioTiempo au = widget.ejercicio;
-      if(true) {
+      if (true) {
         ejstate.setTiempo(au.time);
         aux.setTime(ejstate.getTime);
-      
-      }Historial hist = Historial(
+      }
+      Historial hist = Historial(
           dificultad: ej.dificultad,
           ejercicio: widget.ejercicio.name,
           fecha: DateTime.now(),
@@ -135,7 +133,6 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
           activo: true);
       await database.historialDAO.insertHistorial(hist);
     } else {
-      
       EjercicioRepeticiones ej = aux;
       ej.setReps(ejstate.getReps);
       Historial hist = Historial(
@@ -143,20 +140,31 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
           ejercicio: widget.ejercicio.name,
           fecha: DateTime.now(),
           calorias: widget.ejercicio.calories,
-          tipo:"reps",
+          tipo: "reps",
           duracion: null,
           series: ej.series,
           repeticiones: ej.reps,
           idUser: uid,
           activo: true);
-        database.historialDAO.insertHistorial(hist);
-
+      database.historialDAO.insertHistorial(hist);
     }
-    _navigationService.goBack();
+    try {
+      RecomList r = recomListBox.values.singleWhere((element) {
+        var x = element as RecomList;
+        return x.tipo == widget.ejercicio.grupoprincipal;
+      });
+      var auxhechos = r.hechosTotales;
+      var k = r.key;
+      auxhechos++;
+      recomListBox.putAt(
+          k, RecomList(r.tipo, r.ejercicios, r.hechosLista, auxhechos));
+      r2 = recomListBox.values;
+    } catch (e) {} finally {
+      _navigationService.goBack();
+    }
   }
 
- void sacarEdad ()async {
-
+  void sacarEdad() async {
     DateTime endDate = DateTime.now();
     /*
     final database = Provider.of<AppDatabase>(context, listen: false);
@@ -169,7 +177,8 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
     if(edad == null){
       edad = 50;
     }
-  */var state = context.read<LoginState>();
+  */
+    var state = context.read<LoginState>();
     //state.setDate(_dateTime);
     //print(state.getFNacimiento());
     int edad = endDate.year - state.getFNacimiento().year;
@@ -178,8 +187,6 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
     getHealthRate(edad, context); // Si true es que se ha pasado
     // return  edad;
   }
-
-
 
   Widget widgetEj(Ejercicio ejercicio, EjercicioState ejstate) {
     if (ejercicio is EjercicioTiempo) {
@@ -195,112 +202,117 @@ class _BuildEjercicioState extends State<BuildEjercicio> {
       ejstate.setReps(ejr.reps);
       ejstate.setTipo("R");
       return Container(
-        child: Repcount(ejercicioRepeticiones: ejr,ejstate: ejstate,),
+        child: Repcount(
+          ejercicioRepeticiones: ejr,
+          ejstate: ejstate,
+        ),
       );
     }
-
   }
 
+  _initBox() async {
+    recomListBox = await Hive.openBox('recomlists');
+  }
 
   @override
   Widget build(BuildContext context) {
+    _initBox();
     Ejercicio ej = widget.ejercicio;
     var ejstatus = context.watch<EjercicioState>();
     ejstatus.setEjercicio(ej);
-    return  WillPopScope(
-        onWillPop:() {
-          print('Backbutton pressed (device or appbar button), do whatever you want.');
-          _navigationService.goBack();
-          return Future.value(false);
-        },
-        child: Scaffold(
-          // Widget con app prediseñada, esquema
 
-          appBar: AppBar(
-            leading: BackButton(
-              onPressed:
-                _navigationService.goBack,
-            ),
-            title: Text(widget.ejercicio.name),
-            backgroundColor: Colors.indigo,
+    return WillPopScope(
+      onWillPop: () {
+        print(
+            'Backbutton pressed (device or appbar button), do whatever you want.');
+        _navigationService.goBack();
+        return Future.value(false);
+      },
+      child: Scaffold(
+        // Widget con app prediseñada, esquema
+
+        appBar: AppBar(
+          leading: BackButton(
+            onPressed: _navigationService.goBack,
           ),
-          body: Container(
-              child: Column(
-            children: <Widget>[
-              // MyRitmo(),
-              const SizedBox(
-                height: 20,
-              ),
-              widgetEj(ej, ejstatus),
-
-              const SizedBox(
-                height: 20,
-              ),
-
-              Expanded(
-                flex: 1,
-                child: SingleChildScrollView(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 18,
-                      vertical: 19,
-                    ),
-                    child: Text("Descripcion: " + widget.ejercicio.description,
-                        style: TextStyle(
-                          fontSize: 27,
-                        ))),
-              ),
-              const SizedBox(
-                height: 20,
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  RaisedButton(
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12)),
-                    autofocus: true,
-                    color: Colors.indigo,
-                    onPressed:() => addEjercicioBase(context, ejstatus),
-                    padding: EdgeInsets.all(15.0),
-                    child: Text(
-                      "Hecho",
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 25,
-
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                  Padding(
-                    padding: EdgeInsets.all(16.0),
-                  ),
-                  RaisedButton(
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12)),
-                    autofocus: true,
-                    color: Colors.indigo,
-                    onPressed:() => addEjercicio(context, ejstatus,widget.ejercicio,true),
-                    padding: EdgeInsets.all(15.0),
-                    child: Text(
-                      "Hecho",
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 25,
-
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-
-                ],
-              ),
-              const SizedBox(
-                height: 20,
-              ),
-            ],
-          )),
+          title: Text(widget.ejercicio.name),
+          backgroundColor: Colors.indigo,
         ),
+        body: Container(
+            child: Column(
+          children: <Widget>[
+            // MyRitmo(),
+            const SizedBox(
+              height: 20,
+            ),
+            widgetEj(ej, ejstatus),
+
+            const SizedBox(
+              height: 20,
+            ),
+
+            Expanded(
+              flex: 1,
+              child: SingleChildScrollView(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 18,
+                    vertical: 19,
+                  ),
+                  child: Text("Descripcion: " + widget.ejercicio.description,
+                      style: TextStyle(
+                        fontSize: 27,
+                      ))),
+            ),
+            const SizedBox(
+              height: 20,
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                RaisedButton(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12)),
+                  autofocus: true,
+                  color: Colors.indigo,
+                  onPressed: () => addEjercicioBase(context, ejstatus),
+                  padding: EdgeInsets.all(15.0),
+                  child: Text(
+                    "Hecho",
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 25,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ),
+                Padding(
+                  padding: EdgeInsets.all(16.0),
+                ),
+                RaisedButton(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12)),
+                  autofocus: true,
+                  color: Colors.indigo,
+                  onPressed: () =>
+                      addEjercicio(context, ejstatus, widget.ejercicio, true),
+                  padding: EdgeInsets.all(15.0),
+                  child: Text(
+                    "Hecho Entero",
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 25,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(
+              height: 20,
+            ),
+          ],
+        )),
+      ),
     );
   }
 }
